# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BusinessOS API
  description: |
    AI-powered business operating system with multi-agent orchestration,
    RAG-based context management, and real-time collaboration.
  version: 1.0.0
  contact:
    name: BusinessOS Team
    email: support@businessos.dev
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.businessos.dev
    description: Production server
security:
  - BearerAuth: []
  - SessionCookie: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT bearer token for API requests
    SessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set after login
...
paths:
...
  /osa/generate/{app_id}/stream:
    get:
      tags:
        - OSA
      summary: Stream app generation progress (SSE)
      description: |
        Server-Sent Events endpoint for real-time build progress.
        Use EventSource API in browser to consume events.
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SSE stream of build events
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: build_started
                data: {"app_id":"...","progress_percent":0}

                event: progress
                data: {"progress_percent":25,"status_message":"Analyzing requirements..."}

                event: build_completed
                data: {"app_id":"...","deployment_url":"https://..."}
...